Fix the README.md environment variable names
This updates the port environment variable names to be prefixed with
`PDG_TUTORIALS_`. In addition, this mentions that on some systems, the
wait to healthy state can be up to 15 minutes, and bringing down the
containers via `docker-compose down` will lose any tutorial progress.
Finally, this instructs the user to copy env-template.txt
_before_editing it.