DRIVE-BY: Fix docker login
The previous service credentials to pull Docker images are no longer
valid.

This changes the 02-test-initial-state.sh to no longer login to
Docker, and to instead use the corporate Artifactory Docker registry
mirror. To avoid exposing the mirror URL in the public GitHub
repository, this change relies on the PING_IDENTITY_DEVOPS_REGISTRY GitLab CI
variable being set to the Artifactory Docker registry.